full_stack_app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.3
pandas==2.1.3
python-multipart==0.0.6
psycopg2-binary==2.9.9
langchain==0.0.350
openai==1.3.7
sqlalchemy==2.0.23
pydantic==2.5.2
requests==2.31.0
